# docs/swagger.yaml
basePath: /api
definitions:
  dto.ChangePasswordRequest:
    properties:
      new_password:
        example: newpassword123
        maxLength: 255
        minLength: 8
        type: string
      old_password:
        example: oldpassword123
        minLength: 1
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.CreateResourceRequest:
    properties:
      description:
        example: A reusable sample resource
        maxLength: 2000
        type: string
      name:
        example: Example Resource
        maxLength: 120
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        - archived
        example: active
        type: string
    required:
    - name
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: john@example.com
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOi...
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      first_name:
        example: John
        maxLength: 120
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 120
        type: string
      password:
        example: password123
        maxLength: 255
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - password
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        example: newpassword123
        maxLength: 255
        minLength: 8
        type: string
      token:
        example: abc123
        type: string
    required:
    - new_password
    - token
    type: object
  dto.UpdateProfileRequest:
    properties:
      first_name:
        example: John
        maxLength: 120
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 120
        type: string
    required:
    - first_name
    type: object
  dto.UpdateResourceRequest:
    properties:
      description:
        example: Updated description
        maxLength: 2000
        type: string
      name:
        example: Updated Resource
        maxLength: 120
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        - archived
        example: inactive
        type: string
    type: object
  models.APIResponse:
    properties:
      code:
        example: ""
        type: string
      data: {}
      error:
        example: ""
        type: string
      message:
        example: Success
        type: string
      status:
        example: 200
        type: integer
    type: object
  models.PaginatedResponse:
    properties:
      data: {}
      limit:
        example: 10
        type: integer
      message:
        example: Success
        type: string
      page:
        example: 1
        type: integer
      status:
        example: 200
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
host: localhost:4000
info:
  contact: {}
  description: A production-ready REST API template built with Fiber, GORM, PostgreSQL,
    JWT, SQL migrations, and Scalar docs.
  title: Go Fiber Boilerplate API
  version: "2.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent if email exists
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: User login
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, password, and profile name
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Reset password
      tags:
      - Authentication
  /health:
    get:
      description: Check API and database health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Health check
      tags:
      - Health
  /resources:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List resources
      tags:
      - Resources
    post:
      consumes:
      - application/json
      parameters:
      - description: Resource creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Create resource
      tags:
      - Resources
  /resources/{id}:
    delete:
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete resource
      tags:
      - Resources
    get:
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get resource
      tags:
      - Resources
    put:
      consumes:
      - application/json
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Resource update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update resource
      tags:
      - Resources
  /user/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Invalid current password
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Users
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"